rpi4/mqtt_reader/requirements.txt
paho-mqtt
kuksa-client